requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.32.5
openai>=1.0.0
anthropic>=0.25.0
python-dotenv>=1.0.0
